refactor(Avatar, Inventory): migrate avatar logic from Inventory to Avatar class
