+++ src/Interface/Pages/ChestReward/back.js
@@ -101,8 +101,8 @@ class BackChestReward extends PageBase {
         this.rarityColor = themeManager.GetRariryColors(item.Rarity)[0];
 
         // Configuration du nouvel avatar
-        this.avatarBody = user.inventory.avatar.skin || 'human_00';
-        this.avatarBodyColor = BODY_COLORS[user.inventory.avatar.skinColor] || BODY_COLORS[0];
+        this.avatarBody = user.avatar.avatar.skin || 'human_00';
+        this.avatarBodyColor = BODY_COLORS[user.avatar.avatar.skinColor] || BODY_COLORS[0];
         this.avatarItems = this.getPreviewItems(item);
         this.avatarPosition = dataManager.items.GetContainerSize(item.Slot);
         this.callback = props.args.callback;
@@ -147,7 +147,7 @@ class BackChestReward extends PageBase {
 
         // For 'top' items, also show bottom item (like in avatar editor slots)
         if (item.Slot === 'top') {
-            const bottomStuffID = user.inventory.avatar.bottom;
+            const bottomStuffID = user.avatar.avatar.bottom;
             const bottomStuff = user.inventory.GetStuffByID(bottomStuffID);
             const bottomItemID = bottomStuff ? bottomStuff.ItemID : 'bottom_00';
             baseItems.push({ id: bottomItemID });
